Exception Handling

+++ src/main/java/com/waldstonsantana/marketplace/services/ProductService.java
@@ -3,10 +3,15 @@ package com.waldstonsantana.marketplace.services;
 import com.waldstonsantana.marketplace.dtos.ProductDTO;
 import com.waldstonsantana.marketplace.models.Product;
 import com.waldstonsantana.marketplace.repositories.ProductRepository;
+import com.waldstonsantana.marketplace.services.exceptions.DataBaseException;
+import com.waldstonsantana.marketplace.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -20,7 +25,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Product product = repository.findById(id).get();
+        Product product = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado!"));
         return new ProductDTO(product);
 
     }
@@ -46,17 +52,33 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id,  ProductDTO dto) {
+       try {
+
         Product entity = repository.getReferenceById(id);
         copyDtoToEntity(dto,entity);
 
         entity = repository.save(entity);
 
         return  new ProductDTO(entity);
+       }
+       catch (EntityNotFoundException e) {
+           throw  new ResourceNotFoundException("Recurso não encontrado");
+       }
 
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if(!repository.existsById(id)) {
+            throw  new ResourceNotFoundException("Recurso não encontrado!");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw  new DataBaseException("Falha de integridade referencial");
+        }
+
         repository.deleteById(id);
 
     }
